pyproject.toml
[project]
name="lern_tg_bot"
version="0.0.0"
requires-python = ">=3.8.10,<3.9"
dependencies = [
  "python-dotenv",
  "aiogram",
  "peewee",
]

[tool.black]
line-length = 79

[tool.ruff]
line-length = 79